[fix] coupons : rate 자료형 변경 Double -> Integer #15

+++ coupon/src/main/java/com/haot/coupon/application/dto/request/coupons/CouponCreateRequest.java
@@ -15,7 +15,7 @@ public record CouponCreateRequest(
         DiscountPolicy discountPolicy,
         Double minimumAmount,
         Double maximumAmount,
-        Double discountRate,
+        Integer discountRate,
         Double discountAmount,
         Integer maxQuantity,
         Integer issuedQuantity
